Refactor: V-1.0.3 Organize project structure into data, domain and presentation layers
This commit refactors the project by:

- Moving data models (`FakeData`, `League`, `Player`, `Team`) into a new `data` package and renaming `FakeData` to `LeagueData`.
- Moving UI-related classes (`MainActivity`, `PlayerInfoBottomSheet`, `RankingFragment`) into a new `presentation` package.
- Updating `AndroidManifest.xml` and layout files to reflect the new package structure.
- Removing the static `application` instance from `MyApp.kt`.
- Adding a Kotlin compiler option to silence a warning related to property parameter annotation default target mode.
