+++ app/build.gradle
@@ -14,8 +14,8 @@ android {
         applicationId "ir.miare.androidcodechallenge"
         minSdk 21
         targetSdk 35
-        versionCode 3
-        versionName "1.0.2"
+        versionCode 4
+        versionName "1.0.3"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -32,6 +32,8 @@ android {
     }
     kotlinOptions {
         jvmTarget = '17'
+        freeCompilerArgs = ['-XXLanguage:+PropertyParamAnnotationDefaultTargetMode']
+        // I don't know what consequences i might face with this line in future Just to silent god damn warning:D
     }
 
     buildFeatures {
