+++ app/src/main/java/ir/miare/androidcodechallenge/MyApp.kt
@@ -4,14 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class MyApp : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        application = this
-    }
-
-    companion object {
-        lateinit var application: Application
-    }
-}
+class MyApp : Application()
