+++ app/src/main/java/ir/miare/androidcodechallenge/data/LeagueData.kt
@@ -1,0 +1,8 @@
+package ir.miare.androidcodechallenge.data
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class LeagueData(
+    @JsonProperty("league") var league: League,
+    @JsonProperty("players") var players: List<Player>
+)
