+++ app/src/main/java/ir/miare/androidcodechallenge/data/Team.kt
@@ -1,0 +1,9 @@
+package ir.miare.androidcodechallenge.data
+
+import com.fasterxml.jackson.annotation.JsonProperty
+import java.io.Serializable
+
+data class Team(
+    @JsonProperty("name") val name: String,
+    @JsonProperty("rank") val rank: Int
+) : Serializable
