+++ app/src/main/java/ir/miare/androidcodechallenge/presentation/MainActivity.kt
@@ -1,7 +1,8 @@
-package ir.miare.androidcodechallenge
+package ir.miare.androidcodechallenge.presentation
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import ir.miare.androidcodechallenge.R
 import ir.miare.androidcodechallenge.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
