+++ app/src/main/res/layout/activity_main.xml
@@ -4,16 +4,16 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".presentation.MainActivity">
 
     <LinearLayout
         android:id="@+id/ll_sorting"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:background="@color/indigo_200"
+        android:orientation="vertical"
         android:paddingTop="16dp"
         android:paddingBottom="16dp"
-        android:orientation="vertical"
-        android:background="@color/indigo_200"
         app:layout_constraintTop_toTopOf="parent">
 
         <TextView
@@ -21,51 +21,55 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="center"
-            android:textStyle="bold"
             android:text="Sorting by: "
-            app:layout_constraintTop_toTopOf="parent"/>
+            android:textStyle="bold"
+            app:layout_constraintTop_toTopOf="parent" />
 
         <LinearLayout
             android:id="@+id/ll_sorting_buttons"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_marginLeft="8dp"
             android:layout_marginTop="16dp"
             android:layout_marginRight="8dp"
-            android:layout_marginLeft="8dp"
-            android:orientation="horizontal"
-            android:gravity="center"
-            android:weightSum="3"
             android:dividerPadding="8dp"
+            android:gravity="center"
+            android:orientation="horizontal"
+            android:weightSum="3"
             app:layout_constraintTop_toBottomOf="@id/tv_sorting_title">
 
             <RadioGroup
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content">
+
                 <RadioButton
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:text="Team &amp; league ranking"
                     android:textColor="@color/indigo_700"
-                    android:textStyle="bold"
-                    android:text="Team &amp; league ranking"/>
+                    android:textStyle="bold" />
+
                 <RadioButton
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:text="Most goals scored by a player"
                     android:textColor="@color/indigo_700"
-                    android:textStyle="bold"
-                    android:text="Most goals scored by a player"/>
+                    android:textStyle="bold" />
+
                 <RadioButton
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:text="Average goal per match in a league"
                     android:textColor="@color/indigo_700"
-                    android:textStyle="bold"
-                    android:text="Average goal per match in a league"/>
+                    android:textStyle="bold" />
+
                 <RadioButton
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:textColor="@color/indigo_700"
-                    android:textStyle="bold"
                     android:checked="true"
-                    android:text="None"/>
+                    android:text="None"
+                    android:textColor="@color/indigo_700"
+                    android:textStyle="bold" />
             </RadioGroup>
 
         </LinearLayout>
@@ -76,9 +80,9 @@
         android:id="@+id/fragmentContainer"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/ll_sorting"
-        app:layout_constraintBottom_toBottomOf="parent"/>
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/ll_sorting" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
